support more number data types

+++ src/handler/handler.rs
@@ -33,7 +33,7 @@ pub struct ColumnFilter {
     pub col: String,
     pub filter_type: String,
     pub data_type: String,
-    pub value_int: Option<i64>,
+    pub value_int: Option<i128>,
     pub value_flt: Option<f64>,
     pub value_bol: Option<bool>,
     pub value_str: Option<String>,
